agents/vocabulary_agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
